+++ trlc/ast.py
@@ -130,7 +130,7 @@ class Node(metaclass=ABCMeta):
 
         This can be called for any :class:`Node` or
         :class:`Symbol_Table`, and can be very helpful for debugging
-        or understanding the parse tree. The dump methods will produce
+        or understanding the parse tree. The dump method will produce
         output like this::
 
             Symbol_Table
@@ -342,6 +342,26 @@ class Check(Node):
         # raised is non-fatal.
         self.message   = t_message.value.replace("\n", " ")
         self.extrainfo = extrainfo
+        self._uses_field_access = None
+
+    @property
+    def uses_field_access(self):
+        """Cached test: does this check's expression follow a record/union
+        reference?
+
+        Returns True if any sub-expression of the check expression is a
+        :class:`Field_Access_Expression` whose prefix has a
+        :class:`Record_Type` or :class:`Union_Type` type. Used by
+        the VCG to split checks into Phase A ("at declaration") and
+        Phase B ("after references").
+
+        :return: whether this check dereferences a record/union reference
+        :rtype: bool
+
+        """
+        if self._uses_field_access is None:
+            self._uses_field_access = self.n_expr.uses_field_access()
+        return self._uses_field_access
 
     def dump(self, indent=0):  # pragma: no cover
         # lobster-exclude: Debugging feature
@@ -544,6 +564,22 @@ class Expression(Node, metaclass=ABCMeta):
         """
         assert False, "can_be_null not implemented for %s" % \
             self.__class__.__name__
+
+    def uses_field_access(self):
+        """Test if this expression contains a field access on a record or
+        union reference.
+
+        Returns True if any sub-expression is a
+        :class:`Field_Access_Expression` whose prefix has a
+        :class:`Record_Type` or :class:`Union_Type` type. This is
+        used by the VCG to split checks into "at declaration"
+        (Phase A) and "after references" (Phase B).
+
+        :return: whether this expression follows a record/union reference
+        :rtype: bool
+
+        """
+        return False
 
 
 class Implicit_Null(Expression):
@@ -954,6 +990,9 @@ class Array_Aggregate(Expression):
     def can_be_null(self):
         return False
 
+    def uses_field_access(self):
+        return any(expr.uses_field_access() for expr in self.value)
+
 
 class Tuple_Aggregate(Expression):
     """Instances of a tuple
@@ -1047,6 +1086,10 @@ class Tuple_Aggregate(Expression):
     def can_be_null(self):
         return False
 
+    def uses_field_access(self):
+        return any(expr.uses_field_access()
+                   for expr in self.value.values())
+
 
 class Record_Reference(Expression):
     """Reference to another record object
@@ -1083,7 +1126,7 @@ class Record_Reference(Expression):
         # lobster-exclude: Constructor only declares variables
         assert isinstance(location, Location)
         assert isinstance(name, str)
-        assert isinstance(typ, Record_Type) or typ is None
+        assert isinstance(typ, (Record_Type, Union_Type)) or typ is None
         assert isinstance(package, Package)
         super().__init__(location, typ)
 
@@ -1108,6 +1151,7 @@ class Record_Reference(Expression):
 
     def resolve_references(self, mh):
         # lobster-trace: LRM.References_To_Extensions
+        # lobster-trace: LRM.Union_Type_Minimum_Members
         assert isinstance(mh, Message_Handler)
 
         self.target = self.package.symbols.lookup_direct(
@@ -1117,6 +1161,14 @@ class Record_Reference(Expression):
             required_subclass = Record_Object)
         if self.typ is None:
             self.typ = self.target.n_typ
+        elif isinstance(self.typ, Union_Type):
+            if not self.typ.is_compatible(self.target.n_typ):
+                mh.error(self.location,
+                         "expected reference of type %s,"
+                         " but %s is of type %s" %
+                         (self.typ.name,
+                          self.target.name,
+                          self.target.n_typ.name))
         elif not self.target.n_typ.is_subclass_of(self.typ):
             mh.error(self.location,
                      "expected reference of type %s, but %s is of type %s" %
@@ -1343,6 +1395,9 @@ class Unary_Expression(Expression):
     def can_be_null(self):
         return False
 
+    def uses_field_access(self):
+        return self.n_operand.uses_field_access()
+
 
 class Binary_Expression(Expression):
     """Expression with two operands
@@ -1414,9 +1469,42 @@ class Binary_Expression(Expression):
 
         elif operator in (Binary_Operator.COMP_EQ,
                           Binary_Operator.COMP_NEQ):
+            # lobster-trace: LRM.Union_Type_Equality
+            # lobster-trace: LRM.Union_Type_Equality_Domain
             if (self.n_lhs.typ is None) or (self.n_rhs.typ is None):
                 # We can compary anything to null (including itself)
                 pass
+            elif isinstance(self.n_lhs.typ, Union_Type) or \
+                 isinstance(self.n_rhs.typ, Union_Type):
+                # For union types, we allow comparison if both
+                # sides are record-like (Record_Type or Union_Type)
+                lhs_is_record = isinstance(self.n_lhs.typ,
+                                           (Record_Type, Union_Type))
+                rhs_is_record = isinstance(self.n_rhs.typ,
+                                           (Record_Type, Union_Type))
+                if not (lhs_is_record and rhs_is_record):
+                    mh.error(self.location,
+                             "type mismatch: %s and %s do not match" %
+                             (self.n_lhs.typ.name,
+                              self.n_rhs.typ.name))
+                else:
+                    # Check that there is at least one pair of member
+                    # types (one from each side) where one is a subtype
+                    # of the other. This implements Equality_Domain for
+                    # union types: an unrelated record type is rejected.
+                    lhs_members = (self.n_lhs.typ.types
+                                   if isinstance(self.n_lhs.typ, Union_Type)
+                                   else [self.n_lhs.typ])
+                    rhs_members = (self.n_rhs.typ.types
+                                   if isinstance(self.n_rhs.typ, Union_Type)
+                                   else [self.n_rhs.typ])
+                    if not any(lm.is_subclass_of(rm) or rm.is_subclass_of(lm)
+                               for lm in lhs_members
+                               for rm in rhs_members):
+                        mh.error(self.location,
+                                 "type mismatch: %s and %s do not match" %
+                                 (self.n_lhs.typ.name,
+                                  self.n_rhs.typ.name))
             elif self.n_lhs.typ != self.n_rhs.typ:
                 # Otherwise we can compare anything, as long as the
                 # types match
@@ -1741,31 +1829,49 @@ class Binary_Expression(Expression):
     def can_be_null(self):
         return False
 
+    def uses_field_access(self):
+        return (self.n_lhs.uses_field_access() or
+                self.n_rhs.uses_field_access())
+
 
 class Field_Access_Expression(Expression):
-    """Tuple or Record field access
+    """Tuple, Record, or Union field access
 
     For example in::
 
       foo.bar
       ^1  ^2
 
-    :attribute n_prefix: expression with tuple or record type (see 1)
+    :attribute n_prefix: expression with tuple, record, or union type (see 1)
     :type: Expression
 
-    :attribute n_field: a tuple field to dereference (see 2)
+    :attribute n_field: a field to dereference (see 2)
     :type: Composite_Component
 
+    :attribute is_union_access: True if the prefix is a union type
+    :type: bool
+
+    :attribute is_universal: True if field exists in all union members.
+    Only meaningful when is_union_access is True.
+    :type: bool
+
     """
-    def __init__(self, mh, location, n_prefix, n_field):
+    def __init__(self, mh, location, n_prefix, n_field,
+                 is_union_access=False, is_universal=True):
+        # lobster-trace: LRM.Union_Type_Field_Access
         assert isinstance(mh, Message_Handler)
         assert isinstance(n_prefix, Expression)
         assert isinstance(n_field, Composite_Component)
+        assert isinstance(is_union_access, bool)
+        assert isinstance(is_universal, bool)
         super().__init__(location, n_field.n_typ)
-        self.n_prefix = n_prefix
-        self.n_field  = n_field
+        self.n_prefix        = n_prefix
+        self.n_field         = n_field
+        self.is_union_access = is_union_access
+        self.is_universal    = is_universal
 
-        self.n_prefix.ensure_type(mh, self.n_field.member_of)
+        if not is_union_access:
+            self.n_prefix.ensure_type(mh, self.n_field.member_of)
 
     def dump(self, indent=0):  # pragma: no cover
         # lobster-exclude: Debugging feature
@@ -1788,14 +1894,29 @@ class Field_Access_Expression(Expression):
             mh.error(self.n_prefix.location,
                      "null dereference")
 
+        # lobster-trace: LRM.Union_Type_Partial_Field_Access
+        # lobster-trace: LRM.Union_Type_Partial_Field_Null
+        if self.n_field.name not in v_prefix:
+            return Value(self.location, None, None)
+
         v_field = v_prefix[self.n_field.name]
-        if isinstance(v_field, Implicit_Null):
+        if isinstance(v_field, Expression):
+            # lobster-trace: LRM.Dereference
             return v_field.evaluate(mh, context, gstab)
         else:
             return v_field
 
     def can_be_null(self):
-        return False
+        # A union field access on a partial field (not present in all
+        # member types) evaluates to null at runtime, so we must
+        # report True in that case.
+        return self.is_union_access and not self.is_universal
+
+    def uses_field_access(self):
+        # lobster-trace: LRM.Dereference
+        if isinstance(self.n_prefix.typ, (Record_Type, Union_Type)):
+            return True
+        return self.n_prefix.uses_field_access()
 
 
 class Range_Test(Expression):
@@ -1881,6 +2002,11 @@ class Range_Test(Expression):
     def can_be_null(self):
         return False
 
+    def uses_field_access(self):
+        return (self.n_lhs.uses_field_access() or
+                self.n_lower.uses_field_access() or
+                self.n_upper.uses_field_access())
+
 
 class OneOf_Expression(Expression):
     """OneOf expression
@@ -1932,6 +2058,9 @@ class OneOf_Expression(Expression):
 
     def can_be_null(self):
         return False
+
+    def uses_field_access(self):
+        return any(n_choice.uses_field_access() for n_choice in self.choices)
 
 
 class Action(Node):
@@ -2079,6 +2208,13 @@ class Conditional_Expression(Expression):
         return any(action.n_expr.can_be_null()
                    for action in self.actions)
 
+    def uses_field_access(self):
+        return (any(action.n_cond.uses_field_access() or
+                    action.n_expr.uses_field_access()
+                    for action in self.actions) or
+                (self.else_expr is not None and
+                 self.else_expr.uses_field_access()))
+
 
 class Quantified_Expression(Expression):
     """A quantified expression
@@ -2192,6 +2328,10 @@ class Quantified_Expression(Expression):
 
     def can_be_null(self):
         return False
+
+    def uses_field_access(self):
+        return (self.n_source.uses_field_access() or
+                self.n_expr.uses_field_access())
 
 
 ##############################################################################
@@ -2438,6 +2578,112 @@ class Array_Type(Type):
         return "[%s]" % self.element_type.get_example_value()
 
 
+class Union_Type(Type):
+    # lobster-trace: LRM.union_type
+    # lobster-trace: LRM.Union_Type_Minimum_Members
+    # lobster-trace: LRM.Union_Type_Record_Types_Only
+    """Anonymous union type for record references.
+
+    These are declared implicitly when a record component specifies
+    multiple allowed record types using bracket syntax::
+
+      parent [Systemrequirement, Codebeamerrequirement]
+             ^
+
+    :attribute types: the allowed record types
+    :type: list[Record_Type]
+
+    """
+    def __init__(self, location, types):
+        assert isinstance(types, list)
+        assert len(types) >= 1
+        assert all(isinstance(t, Record_Type) for t in types)
+        name = "[%s]" % ", ".join(t.name for t in types)
+        super().__init__(name, location)
+        self.types = types
+        self._field_map = None
+
+    def get_field_map(self):
+        # lobster-trace: LRM.Union_Type_Field_Access
+        """Compute accessible fields across all union members.
+
+        Returns a dict mapping field name to a dict with keys:
+
+        * ``component``: a representative Composite_Component
+        * ``n_typ``: the field type (None if conflicting)
+        * ``count``: how many member types have this field
+        * ``total``: total number of member types
+        * ``optional_in_any``: True if optional in at least one member
+
+        :rtype: dict[str, dict]
+        """
+        if self._field_map is not None:
+            return self._field_map
+
+        field_map = {}
+        for record_type in self.types:
+            seen_in_type = set()
+            for comp in record_type.all_components():
+                if comp.name in seen_in_type:
+                    continue
+                seen_in_type.add(comp.name)
+                if comp.name not in field_map:
+                    field_map[comp.name] = {
+                        "component"       : comp,
+                        "n_typ"           : comp.n_typ,
+                        "count"           : 1,
+                        "total"           : len(self.types),
+                        "optional_in_any" : comp.optional,
+                    }
+                else:
+                    info = field_map[comp.name]
+                    info["count"] += 1
+                    # Type identity (is) is correct here:
+                    # non-union type objects are structural
+                    # singletons in the symbol table, so
+                    # identity comparison is both correct and
+                    # cheap.
+                    if info["n_typ"] is not comp.n_typ:
+                        info["n_typ"] = None  # type conflict
+                    if comp.optional:
+                        info["optional_in_any"] = True
+
+        self._field_map = field_map
+        return self._field_map
+
+    def dump(self, indent=0):  # pragma: no cover
+        # lobster-exclude: Debugging feature
+        self.write_indent(indent, "Union_Type")
+        for t in self.types:
+            self.write_indent(indent + 1, t.name)
+
+    def perform_type_checks(self, mh, value, gstab):
+        # Union types have no checks of their own; type validation
+        # happens in Record_Reference.resolve_references() via
+        # is_compatible().  Returning True unconditionally is
+        # intentional.
+        assert isinstance(mh, Message_Handler)
+        assert isinstance(value, Expression)
+        assert isinstance(gstab, Symbol_Table)
+        return True
+
+    def is_compatible(self, record_type):
+        """Test if the given record type is accepted by this union.
+
+        :param record_type: type to check
+        :type record_type: Record_Type
+
+        :returns: true if the type is or extends one of the union members
+        :rtype: bool
+        """
+        assert isinstance(record_type, Record_Type)
+        return any(record_type.is_subclass_of(t) for t in self.types)
+
+    def get_example_value(self):
+        # lobster-exclude: utility method
+        return "%s_instance" % self.types[0].name
+
+
 class Builtin_Integer(Builtin_Numeric_Type):
     # lobster-trace: LRM.Builtin_Types
     # lobster-trace: LRM.Integer_Values
@@ -2547,7 +2793,7 @@ class Composite_Type(Concrete_Type, metaclass=ABCMeta):
     :attribute description: user-supplied description of the type or None
     :type: str
 
-    :attribute checks: used-defined checks for this type (excluding \
+    :attribute checks: user-defined checks for this type (excluding \
       inherited checks)
     :type: list[Check]
 
@@ -3302,7 +3548,7 @@ class Symbol_Table:
         """Retrieve an object from the table assuming its there
 
         This is intended for the API specifically where you want to
-        e.g. find some used-defined types you know are there.
+        e.g. find some user-defined types you know are there.
 
         :param mh: The message handler to use
         :type mh: Message_Handler